+++ src/axom/inlet/Function.hpp
@@ -1,0 +1,320 @@
+// Copyright (c) 2017-2020, Lawrence Livermore National Security, LLC and
+// other Axom Project Developers. See the top-level COPYRIGHT file for details.
+//
+// SPDX-License-Identifier: (BSD-3-Clause)
+
+/*!
+ *******************************************************************************
+ * \file Function.hpp
+ *
+ * \brief This file contains the class definition of Inlet's Function class.
+ *******************************************************************************
+ */
+
+#ifndef INLET_FUNCTION_HPP
+#define INLET_FUNCTION_HPP
+
+#include <memory>
+#include <functional>
+#include <tuple>
+#include <type_traits>
+
+#include "fmt/fmt.hpp"
+
+#include "axom/sidre.hpp"
+#include "axom/primal/geometry/Vector.hpp"
+
+#include "axom/inlet/Field.hpp"
+#include "axom/inlet/Verifiable.hpp"
+#include "axom/inlet/inlet_utils.hpp"
+
+#include "axom/sidre.hpp"
+
+namespace axom
+{
+namespace inlet
+{
+/*!
+ *******************************************************************************
+ * \brief The tags used to describe function signatures in the input file
+ * 
+ * \note Additions to this enumeration should be propagated to LuaReader
+ * and func_signature_tuples defined below (the mapping from enum to types)
+ * 
+ * \note Vec3D corresponds to a three-dimensional vector, Double corresponds to
+ * a floating-point scalar
+ * 
+ * \note A two-dimensional vector was intentionally excluded for simplicity as
+ * a 3D vector can be used in its place (with the third component being empty/zero)
+ *******************************************************************************
+ */
+enum class FunctionType
+{
+  Vec3D,
+  Double
+};
+
+namespace detail
+{
+/*!
+ *******************************************************************************
+ * \class inlet_function_arg_type
+ *
+ * \brief A type trait for modifying function argument types to enforce const
+ * correctness and to avoid copies
+ * \tparam Arg The argument type
+ * 
+ * Maps a type "Arg" to "const Arg&" if that type does not satisfy std::is_arithmetic,
+ * which matches Inlet primitive types bool, int, and double, but not vectors or
+ * std::strings - though only vectors and doubles are currently supported as arg
+ * types
+ *******************************************************************************
+ */
+template <typename Arg>
+struct inlet_function_arg_type
+{
+  using type = typename std::conditional<
+    std::is_arithmetic<Arg>::value,
+    Arg,
+    typename std::add_lvalue_reference<typename std::add_const<Arg>::type>::type>::type;
+};
+
+/*!
+ *******************************************************************************
+ * \class cleanup_function_signature
+ *
+ * \brief Takes a Ret(Args...) signature and applies cvref qualifiers to applicable
+ * arguments
+ * 
+ * \tparam Ret The function's return type
+ * \tparam Args... The function's arguments
+ * 
+ * Designed to be used with user-facing retrieval functions so the user is not 
+ * required to specify the qualifiers in something like a get<std::function<Ret(Args...)>
+ *******************************************************************************
+ */
+template <typename>
+struct cleanup_function_signature;
+
+template <typename Ret, typename... Args>
+struct cleanup_function_signature<Ret(Args...)>
+{
+  using type = Ret(typename inlet_function_arg_type<Args>::type...);
+};
+
+}  // end namespace detail
+
+/*!
+ *******************************************************************************
+ * \class FunctionWrapper
+ *
+ * \brief A sum type for callables with arbitrary signature
+ * \tparam FunctionTypes The types of supported functions
+ * 
+ * This provides an interface not templated on a specific function signature
+ * for uniform retrieval through the Reader interface
+ *******************************************************************************
+ */
+template <typename... FunctionTypes>
+class FunctionWrapper
+{
+public:
+  /*!
+   *******************************************************************************
+   * \brief Primary constructor, initializes a member of the "variant"
+   * 
+   * \param [in] func The function to initialize with
+   * \tparam FuncType The function's signature
+   * 
+   * \note "Empty" functions are allowable and result in the object comparing
+   * false when converted to bool
+   *******************************************************************************
+   */
+  template <typename FuncType>
+  FunctionWrapper(std::function<FuncType>&& func)
+  {
+    m_function_valid = static_cast<bool>(func);
+    std::get<std::unique_ptr<std::function<FuncType>>>(m_funcs) =
+      cpp11_compat::make_unique<std::function<FuncType>>(std::move(func));
+  }
+
+  FunctionWrapper() = default;
+  FunctionWrapper(FunctionWrapper&& other) = default;
+
+  /*!
+   *******************************************************************************
+   * \brief Calls the function
+   * 
+   * \param [in] args The parameter pack for the function's arguments
+   * \tparam Ret The user-specified return type, needed to fully disambiguate the
+   * function to call
+   * \tparam Args The types of the user-specified arguments, deduced automatically
+   * 
+   * \return The function's result
+   *******************************************************************************
+   */
+  template <typename Ret, typename... Args>
+  Ret call(Args&&... args) const
+  {
+    const auto& ptr = std::get<std::unique_ptr<
+      std::function<Ret(typename detail::inlet_function_arg_type<Args>::type...)>>>(
+      m_funcs);
+    SLIC_ERROR_IF(
+      !m_function_valid || !ptr || !(*ptr),
+      fmt::format("[Inlet] Function '{0}' with requested type does not exist",
+                  m_name));
+
+    const auto& func = *ptr;
+    return func(
+      std::forward<typename detail::inlet_function_arg_type<Args>::type>(args)...);
+  }
+
+  template <typename FuncType>
+  std::function<FuncType> get() const
+  {
+    const auto& ptr = std::get<std::unique_ptr<
+      std::function<typename detail::cleanup_function_signature<FuncType>::type>>>(
+      m_funcs);
+    SLIC_ERROR_IF(
+      !ptr,
+      fmt::format("[Inlet] Function '{0}' with requested type does not exist",
+                  m_name));
+    return *ptr;
+  }
+
+  /*!
+   *******************************************************************************
+   * \brief Checks whether the function exists
+   *******************************************************************************
+   */
+  explicit operator bool() const { return m_function_valid; }
+
+  /*!
+   *******************************************************************************
+   * \brief Sets the function's name
+   * 
+   * \note Needs to be separate from constructor to allow the compiler to deduce
+   * template arguments correctly
+   *******************************************************************************
+   */
+  void setName(std::string&& name) { m_name = std::move(name); }
+
+private:
+  // This is on the heap to reduce size - each pointer is only 8 bytes vs 32 bytes
+  // for a std::function, and it is guaranteed that only one of the pointers will
+  // actually point to something
+  std::tuple<std::unique_ptr<std::function<FunctionTypes>>...> m_funcs;
+  bool m_function_valid = false;
+  std::string m_name;
+};
+
+using FunctionVariant =
+  FunctionWrapper<double(const primal::Vector3D&),
+                  primal::Vector3D(const primal::Vector3D&)>;
+
+class Function : public Verifiable<Function>
+{
+public:
+  Function(axom::sidre::Group* sidreGroup,
+           axom::sidre::Group* root,
+           FunctionVariant&& func,
+           bool docEnabled = true)
+    : m_sidreGroup(sidreGroup)
+    , m_sidreRootGroup(root)
+    , m_docEnabled(docEnabled)
+    , m_func(std::move(func))
+  {
+    m_func.setName(name());
+  }
+
+  /*!
+   *****************************************************************************
+   * \brief Returns a function of requested type
+   * 
+   * \return The value
+   * \tparam T The type to retrieve
+   *****************************************************************************
+   */
+  template <typename FuncType>
+  std::function<FuncType> get() const
+  {
+    return m_func.get<FuncType>();
+  }
+
+  template <typename Ret, typename... Args>
+  Ret call(Args&&... args) const
+  {
+    return m_func.call<Ret>(std::forward<Args>(args)...);
+  }
+
+  /*!
+   *******************************************************************************
+   * \brief Checks whether the function exists
+   *******************************************************************************
+   */
+  operator bool() const { return static_cast<bool>(m_func); }
+
+  /*!
+   *****************************************************************************
+   * \return The full name for this Function.
+   *****************************************************************************
+  */
+  std::string name() const;
+
+  /*!
+   *****************************************************************************
+   * \brief This will be called by Inlet::verify to verify the contents of this
+   *  Function
+   *****************************************************************************
+   */
+  bool verify() const;
+
+  /*!
+   *****************************************************************************
+   * \brief Set the required status of this Table.
+   *
+   * Set whether this Table is required, or not, to be in the input file.
+   * The default behavior is to not be required.
+   *
+   * \param [in] isRequired Boolean value of whether Table is required
+   *
+   * \return Reference to this instance of Table
+   *****************************************************************************
+   */
+  Function& required(bool isRequired = true);
+
+  /*!
+   *****************************************************************************
+   * \brief Return the required status of this Table.
+   *
+   * Return that this Function is required, or not, to be in the input file.
+   * The default behavior is to not be required.
+   *
+   * \return Boolean value of whether this Function is required
+   *****************************************************************************
+   */
+  bool isRequired() const;
+
+  /*!
+   *****************************************************************************
+   * \brief Registers the function object that will verify this function
+   * during the verification stage.
+   * 
+   * \param [in] The function object that will be called by Table::verify().
+   *****************************************************************************
+  */
+  Function& registerVerifier(std::function<bool(const Function&)> lambda);
+
+private:
+  // This function's sidre group
+  axom::sidre::Group* m_sidreGroup = nullptr;
+  axom::sidre::Group* m_sidreRootGroup = nullptr;
+  bool m_docEnabled;
+  std::function<bool(const Function&)> m_verifier;
+  FunctionVariant m_func;
+};
+
+}  // end namespace inlet
+}  // end namespace axom
+
+#endif  // INLET_FUNCTION_HPP