+++ src/axom/inlet/tests/inlet_Inlet.cpp
@@ -1323,10 +1323,10 @@ TEST(inletArrays, getArray)
     "              arr4 = { [12] = 2.4 } }";
   auto inlet = createBasicInlet(&ds, testString);
 
-  inlet.getGlobalTable().addIntArray("luaArrays/arr1");
-  inlet.getGlobalTable().addBoolArray("luaArrays/arr2");
-  inlet.getGlobalTable().addStringArray("luaArrays/arr3");
-  inlet.getGlobalTable().addDoubleArray("luaArrays/arr4");
+  inlet.addIntArray("luaArrays/arr1");
+  inlet.addBoolArray("luaArrays/arr2");
+  inlet.addStringArray("luaArrays/arr3");
+  inlet.addDoubleArray("luaArrays/arr4");
 
   std::unordered_map<int, int> expectedInts {{1, 4}};
   std::unordered_map<int, bool> expectedBools {{4, true}, {8, false}};
@@ -1360,8 +1360,7 @@ TEST(inletArrays, inletArraysInSidre)
 
   inlet.addIntArray("luaArrays/arr1");
 
-  auto group = inlet.getGlobalTable().sidreGroup()->getGroup(
-    "luaArrays/arr1/_inlet_array");
+  auto group = inlet.sidreGroup()->getGroup("luaArrays/arr1/_inlet_container");
   auto idx = group->getGroup("1");
   EXPECT_TRUE(idx);
   int val = idx->getView("value")->getScalar();
@@ -1382,8 +1381,8 @@ TEST(inletArrays, inletArraysInSidre)
   val = idx->getView("value")->getScalar();
   EXPECT_EQ(val, 6);
 
-  inlet.getGlobalTable().addBoolArray("luaArrays/arr2");
-  group = inlet.getTable("luaArrays/arr2/_inlet_array").sidreGroup();
+  inlet.addBoolArray("luaArrays/arr2");
+  group = inlet.getTable("luaArrays/arr2/_inlet_container").sidreGroup();
 
   idx = group->getGroup("4");
   EXPECT_TRUE(idx);
@@ -1395,8 +1394,8 @@ TEST(inletArrays, inletArraysInSidre)
   boolVal = idx->getView("value")->getScalar();
   EXPECT_EQ(boolVal, 0);
 
-  inlet.getGlobalTable().addStringArray("luaArrays/arr3");
-  group = inlet.getTable("luaArrays/arr3/_inlet_array").sidreGroup();
+  inlet.addStringArray("luaArrays/arr3");
+  group = inlet.getTable("luaArrays/arr3/_inlet_container").sidreGroup();
 
   idx = group->getGroup("33");
   EXPECT_TRUE(idx);
@@ -1408,8 +1407,8 @@ TEST(inletArrays, inletArraysInSidre)
   str = idx->getView("value")->getString();
   EXPECT_EQ(str, "bye");
 
-  inlet.getGlobalTable().addDoubleArray("luaArrays/arr4");
-  group = inlet.getTable("luaArrays/arr4/_inlet_array").sidreGroup();
+  inlet.addDoubleArray("luaArrays/arr4");
+  group = inlet.getTable("luaArrays/arr4/_inlet_container").sidreGroup();
 
   idx = group->getGroup("12");
   EXPECT_TRUE(idx);